added user roles

+++ app/Http/Controllers/ProjectController.php
@@ -13,6 +13,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Mail;
+use Illuminate\Validation\Rule;
 
 class ProjectController extends Controller
 {
@@ -21,83 +22,155 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        $query = Project::query();
+        $user = auth()->user();
 
-        $sortField = request('sort_field', 'created_at');
-        $sortDirection = request('sort_direction', 'desc');
+        if ($user->role === 'client') {
 
-        if (request('project_name')) {
-            $query->where("project_name", "like", "%" . request("project_name") . "%");
+            $query = Project::query();
+
+            $sortField = request('sort_field', 'created_at');
+            $sortDirection = request('sort_direction', 'desc');
+
+            if (request('project_name')) {
+                $query->where('project_name', 'like', '%' . request('project_name') . '%');
+            }
+            if (request('status')) {
+                $query->where('status', request('status'));
+            }
+
+            $projects = Project::where('user_id', $user->id)->get();
+
+            return inertia('Client/Project/Index', ['projects' => $projects]); 
         }
-        if (request("status")) {
-            $query->where("status", request("status"));
+
+        if ($user->role === 'admin') {
+
+            $query = Project::query();
+
+            $sortField = request('sort_field', 'created_at');
+            $sortDirection = request('sort_direction', 'desc');
+
+            if (request('project_name')) {
+                $query->where('project_name', 'like', '%' . request('project_name') . '%');
+            }
+            if (request('status')) {
+                $query->where('status', request('status'));
+            }
+
+            $projects = $query->orderBy($sortField, $sortDirection)
+                ->paginate(10)
+                ->onEachSide(1);
+
+            return inertia("Project/Index", [
+                'projects' => ProjectResource::collection($projects),
+                'queryParams' => request()->query() ?: null,
+                'success' => session('success'),
+            ]);
         }
-
-        $projects = $query->orderBy($sortField, $sortDirection)
-            ->paginate(10)
-            ->onEachSide(1);
-
-        return inertia("Project/Index", [
-            "projects" => ProjectResource::collection($projects),
-            'queryParams' => request()->query() ?: null,
-            'success' => session('success'),
-        ]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create($id)
+    public function create()
     {
-        $client = User::where('id', $id)->firstOrFail();
-        $user_id = $client->id;
-        return inertia('Project/Create', [
-            'user_id' => $user_id
-        ]);
+        $user = auth()->user();
+
+        if ($user->role === 'client') {
+
+            return abort(404);
+        }
+
+        if ($user->role === 'admin') {
+
+            return inertia('Project/Create');
+        }
+    }
+
+        /**
+     * Show the form for creating a new resource.
+     */
+    public function create_with_id($id)
+    {
+        $user = auth()->user();
+
+        if ($user->role === 'client') {
+
+            return abort(404);
+        }
+
+        if ($user->role === 'admin') {
+
+            $user_find = User::where('id', $id)->firstOrFail();
+            return inertia('Project/CreateWithId', [
+                'user_id' => $user_find->id
+            ]);
+        }
     }
 
     /**
      * Store a newly created resource in storage.
      */
     // public function store(StoreProjectRequest $request, $id) {
-        public function store(Request $request, Project $project) {
+    public function store(Request $request, Project $project) 
+    {
+        $user = auth()->user();
 
-        // $data = $request->validated();
-        // $project->users()->create({[
-        // ]}) 
-        // $user = User::where('id', $user)->firstOrFail();
-        // dd($request->user()->name);
-        // $id = '2';
+        if ($user->role === 'client') {
 
-        $data['user_id'] = $request->user_id;
-        // $data['user_id'] = $request->input($user->id);
-        // $data['client_name'] = $request->name;
-        // $data['client_email'] = $request->email;
-        $data['project_name'] = $request->project_name;
-        $data['status'] = $request->status;
-        
-        /** @var $image \Illuminate\Http\UploadedFile */
-        $image = $data['image'] ?? null;
-        $data['created_by'] = '2';
-        $data['updated_by'] = Auth::id();
-        if ($image) {
-            $data['image_path'] = $image->store('project/' . Str::random(), 'public');
+            return abort(404);
         }
-        Project::create($data);
 
-        $message = 'This is a test email.';
-        
-        Mail::send('email', [
-            'name' => $request->get('client_name'),
-            'email' => $request->get('client_email'), ],
-            function ($message) {
+        if ($user->role === 'admin') {
+
+            $request->validate([
+                'project_name' => ['required', 'max:255'],
+                // 'client_name' => ['required', 'max:255'],
+                // 'client_email' => ['required', 'max:255'],
+                'image' => ['nullable', 'image'],
+                'description' => ['nullable', 'string'],
+                'start_date' => ['nullable', 'date'],
+                'due_date' => ['nullable', 'date'],
+                'status' => ['required', Rule::in(['pending', 'in_progress', 'completed'])]
+            ]);
+
+            // $data = $request->validated();
+            // $project->users()->create({[
+            // ]}) 
+            // $user = User::where('id', $user)->firstOrFail();
+            // dd($request->user()->name);
+            // $id = '2';
+
+            $data['user_id'] = $request->user_id;
+            // $data['user_id'] = $request->input($user->id);
+            // $data['client_name'] = $request->name;
+            // $data['client_email'] = $request->email;
+            $data['project_name'] = $request->project_name;
+            $data['status'] = $request->status;
+            
+            /** @var $image \Illuminate\Http\UploadedFile */
+            $image = $data['image'] ?? null;
+            $data['created_by'] = '2';
+            $data['updated_by'] = Auth::id();
+            if ($image) {
+                $data['image_path'] = $image->store('project/' . Str::random(), 'public');
+            }
+            Project::create($data);
+
+            $message = 'This is a test email.';
+            
+            Mail::send('email', [
+                'name' => $request->get('client_name'),
+                'email' => $request->get('client_email'), ],
+                function ($message) {
                     $message->from('admin@project_management_laravel_react.com');
                     $message->to('youremail@your_domain', 'Your Name')
                     ->subject('New Project "' . 'Project Name' . '" Created.');
-        });
+            });
 
-        return to_route('project.index')
-            ->with('success', 'Project was created');
+            return to_route('project.index')
+                ->with('success', 'Project was created');
+        }
     }
 
     /**
@@ -105,36 +178,54 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        $query = $project->tasks();
+        $user = auth()->user();
 
-        $sortField = request("sort_field", 'created_at');
-        $sortDirection = request("sort_direction", "desc");
+        if ($user->role === 'client') {
 
-        if (request("project_name")) {
-            $query->where("project_name", "like", "%" . request("project_name") . "%");
-        }
-        if (request("status")) {
-            $query->where("status", request("status"));
-        }
+            if ($project->user_id === $user->id ) {
 
-        $tasks = $query->orderBy($sortField, $sortDirection)
-            ->paginate(10)
-            ->onEachSide(1);
-
-            // Display tasks table if there are tasks to show
-            if(count($tasks) <= 0) {
-                $if_tasks = false;
-            } elseif (count($tasks) > 0) {
-                $if_tasks = true;
+                $project = Project::where('id', $project->id)->firstOrFail();
+            
+                return inertia('Client/Project/Show', [
+                    'project' => $project
+                ]);
             }
 
-        return inertia('Project/Show', [
-            'if_tasks' => $if_tasks,
-            'project' => new ProjectResource($project),
-            'tasks' => TaskResource::collection($tasks),
-            'queryParams' => request()->query() ?: null,
-            'success' => session('success'),
-        ]);
+            return abort(404);
+        }
+            
+        if ($user->role === 'admin') {   
+            $query = $project->tasks();
+
+            $sortField = request("sort_field", 'created_at');
+            $sortDirection = request("sort_direction", "desc");
+
+            if (request("project_name")) {
+                $query->where("project_name", "like", "%" . request("project_name") . "%");
+            }
+            if (request("status")) {
+                $query->where("status", request("status"));
+            }
+
+            $tasks = $query->orderBy($sortField, $sortDirection)
+                ->paginate(10)
+                ->onEachSide(1);
+
+                // Display tasks table if there are tasks to show
+                if(count($tasks) <= 0) {
+                    $if_tasks = false;
+                } elseif (count($tasks) > 0) {
+                    $if_tasks = true;
+                }
+
+            return inertia('Project/Show', [
+                'if_tasks' => $if_tasks,
+                'project' => new ProjectResource($project),
+                'tasks' => TaskResource::collection($tasks),
+                'queryParams' => request()->query() ?: null,
+                'success' => session('success'),
+            ]);
+        }
     }
 
     /**
@@ -142,9 +233,19 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
-        return inertia('Project/Edit', [
-            'project' => new ProjectResource($project),
-        ]);
+        $user = auth()->user();
+
+        if ($user->role === 'client') {
+
+            return abort(404);
+        }
+
+        if ($user->role === 'admin') {
+
+            return inertia('Project/Edit', [
+                'project' => new ProjectResource($project),
+            ]);
+        }
     }
 
     /**
@@ -152,19 +253,29 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
-        $data = $request->validated();
-        $image = $data['image'] ?? null;
-        $data['updated_by'] = Auth::id();
-        if ($image) {
-            if ($project->image_path) {
-                Storage::disk('public')->deleteDirectory(dirname($project->image_path));
-            }
-            $data['image_path'] = $image->store('project/' . Str::random(), 'public');
-        }
-        $project->update($data);
+        $user = auth()->user();
 
-        return to_route('project.index')
-            ->with('success', "Project \"$project->name\" was updated");
+        if ($user->role === 'client') {
+
+            return abort(404);
+        }
+
+        if ($user->role === 'admin') {
+
+            $data = $request->validated();
+            $image = $data['image'] ?? null;
+            $data['updated_by'] = Auth::id();
+            if ($image) {
+                if ($project->image_path) {
+                    Storage::disk('public')->deleteDirectory(dirname($project->image_path));
+                }
+                $data['image_path'] = $image->store('project/' . Str::random(), 'public');
+            }
+            $project->update($data);
+
+            return to_route('project.index')
+                ->with('success', "Project \"$project->name\" was updated");
+        }
     }
 
     /**
@@ -172,12 +283,22 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        $name = $project->name;
-        $project->delete();
-        if ($project->image_path) {
-            Storage::disk('public')->deleteDirectory(dirname($project->image_path));
+        $user = auth()->user();
+
+        if ($user->role === 'client') {
+
+            return abort(404);
         }
-        return to_route('project.index')
-            ->with('success', "Project \"$name\" was deleted");
+
+        if ($user->role === 'admin') {
+
+            $name = $project->name;
+            $project->delete();
+            if ($project->image_path) {
+                Storage::disk('public')->deleteDirectory(dirname($project->image_path));
+            }
+            return to_route('project.index')
+                ->with('success', "Project \"$name\" was deleted");
+        }
     }
 }
